requirements.txt
djangorestframework==3.8
Django==2.1
